Create PythonCode
This is code with functions at the top and then the main body at the bottom for a poker game.
Where cars are selected at random, the user then gets shown 1st card and is asked to guess if the 2nd card will be higher or lower.
Through all this, the user has 10 rounds. If the user fails the amount bet will be taken away from the current amount the amount bet will be added. There are multiple constraints in place 
-> Users can not bet more than they currently have 
-> If the user loses all money game is terminated 
-> If the user get up to 500 or more max amount the game terminates 
-> In every circumstance a message is printed.